BankUnreconciled Service created

+++ ArmsModels/BaseModels/User/DataAuthorizationModel.cs
@@ -10,7 +10,7 @@ namespace ArmsModels.BaseModels
 
     public class DataAuthorizationModel
     {
-        public int? ID { get; set; }
+        public long? ID { get; set; }
         [Required]
         public int? DocumentID { get; set; }
         [Required]
